+++ internal/providers/localcontainer/main_test.go
@@ -1,0 +1,12 @@
+package localcontainer
+
+import (
+	"os"
+	"testing"
+
+	"github.com/openclaw/crabbox/internal/testutil"
+)
+
+func TestMain(m *testing.M) {
+	os.Exit(testutil.RunWithIsolatedUserDirs(m))
+}